hotfix recettes + detail UX

+++ api/recettes/fav_actions.php
@@ -6,21 +6,21 @@
     require_once(__DIR__ . '/../../db/db_connect.php');
     require_once(__DIR__ . '/../../lib/fav_utils.php');
 
-    if(isset($_GET['id_user']) && isset($_GET['action']) && isset($_GET['id_recette'])){
+    if(isset($_GET['id_user']) && isset($_GET['action']) && isset($_GET['id_recette']) && isset($_GET['owner'])){
         $action = $_GET['action'];
         $user_id = $_GET['id_user'];
         $recette_id = $_GET['id_recette'];
-        if($action == 'delete'){
-            deleteRecette($conn, $recette_id);
-            header('Location: ../../public/recettes.php?status=delSuccess');
-        }else if($action == 'addFav'){
+        $owner = $_GET['owner'];
+        if($action == 'addFav'){
             addRecetteFav($conn, $user_id, $recette_id);
             addUpVote($conn, $recette_id);
-            header('Location: ../../public/recettes.php?status=addSuccess');
+            header("Location: https://l1.dptinfo-usmb.fr/~grp9/public/profil.php?owner=$owner&id_recette=$recette_id");
+            exit;
         }else if($action == 'remFav'){
             deleteRecetteFav($conn, $recette_id, $user_id,);
             delUpVote($conn, $recette_id);
-            header('Location: ../../public/recettes.php?status=delFavSuccess');
+            header("Location: https://l1.dptinfo-usmb.fr/~grp9/public/profil.php?owner=$owner&id_recette=$recette_id");
+            exit;
         }else if($action == 'isFav'){
             $res = isFav($conn, $user_id, $recette_id);
             $tab = [];
@@ -28,6 +28,30 @@
             echo(json_encode($tab));
         }
 
+    }else if(isset($_GET['id_user']) && isset($_GET['action']) && isset($_GET['id_recette']) && !isset($_GET['owner'])){
+        $action = $_GET['action'];
+        $user_id = $_GET['id_user'];
+        $recette_id = $_GET['id_recette'];
+        if($action == 'delete'){
+            deleteRecette($conn, $recette_id);
+            header("Location: https://l1.dptinfo-usmb.fr/~grp9/public/recettes.php");
+            exit;
+        }else if($action == 'addFav'){
+            addRecetteFav($conn, $user_id, $recette_id);
+            addUpVote($conn, $recette_id);
+            header("Location: https://l1.dptinfo-usmb.fr/~grp9/public/recettes.php");
+            exit;
+        }else if($action == 'remFav'){
+            deleteRecetteFav($conn, $recette_id, $user_id,);
+            delUpVote($conn, $recette_id);
+            header("Location: https://l1.dptinfo-usmb.fr/~grp9/public/recettes.php");
+            exit;
+        }else if($action == 'isFav'){
+            $res = isFav($conn, $user_id, $recette_id);
+            $tab = [];
+            $tab['fav'] = $res;
+            echo(json_encode($tab));
+        }
     }
